clinker-core: extract streaming fused path into streaming.rs
Move the streaming-fusion analysis and the streaming-output writer
thread out of `executor/mod.rs` into a new `streaming.rs`:
`compute_merge_interleave_fused_sources`, `compute_transform_fused_sources`,
`StreamClass`, `classify_stream_nodes`, `streaming_output_producer`,
`compute_streaming_output_specs`, `StreamingOutputSpec`,
`StreamingOutputTaskOutput`, and `streaming_output`, plus the
`has_single_outgoing` plan helper they share.

`compute_streaming_output_specs`, `streaming_output`, and
`StreamingOutputTaskOutput::fold_into` are widened from file-private to
`pub(super)` so the executor entry point reaches them across the new
module boundary; `streaming_output_producer` and `has_single_outgoing`
stay private (only the streaming module consults them). The `pub(crate)`
classifiers and `StreamingOutputTaskOutput` are re-exported from
`mod.rs`, so `executor::classify_stream_nodes` / `executor::StreamClass`
(the `--explain` buffer-class annotation) and
`executor::StreamingOutputTaskOutput` (the dispatcher's join shape)
resolve unchanged.

Pure relocation: no symbol renames, no public-API change, every existing
test passes unmodified.

Closes #260
